skip pyqir tests if not installed
