+++ requirements-dev.txt
@@ -5,6 +5,7 @@ pennylane<0.40
 pyquil>=4.4; python_version < "3.12"
 pytket>=1.31; python_version >= "3.10"
 qiskit>=0.44,<1.3
+cudaq>=0.9.0
 
 # transpiler extras
 ply>=3.6
